Add files via upload
The issue that appears in the code is that it prints out every possibility for 4 molecule combinations.

I believe that the problem arises from the clusters.py[line 333]/recursive.py[179 line] files. 
The algorithm that we had written to check the possibility of the molecules connecting, for 5 molecule structures, relies on the connectivity matrix storing the distance between the molecules. The connnectivity matrix was modified to directly take in the length between molecules and the  DFS was modified to compare the number taken in from the connectivity matrix with that of the Rcutoff, which is the same as with how the connectivity matrix was before, just checking at a different point to see if the distance between the molecules is <Rcutoff. This process had worked in the unmerged code, so I am not sure where this issue is arising from.
It said in the notes:  # DO NOT WRITE NUMBERS MORE THAN 9 INTO CONNECTIVITY TABLE, so I am not sure if this problem is arising due to this, as I wasn